% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{calculate_aggregate_stats}
\alias{calculate_aggregate_stats}
\title{Create overview stats over the aggregate mental model}
\usage{
calculate_aggregate_stats(edgelist)
}
\arguments{
\item{edgelist}{Likely a raw edgelist data frame generated from parse_mtools_csv.}
}
\value{
A data frame with the following stats:
- weighted betwenness
- weighted indegree
- weighted outdegree
- weighted total degree
}
\description{
Create overview stats over the aggregate mental model
}
