% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{calculate_descriptive_statistics}
\alias{calculate_descriptive_statistics}
\title{Create overview stats for a mental model}
\usage{
calculate_descriptive_statistics(mentalmodel)
}
\arguments{
\item{mentalmodel}{A mtoolr mental model object}
}
\value{
A tibble with the following stats:

- weighted betwenness
- weighted indegree
- weighted outdegree
- weighted total degree

If a aggregated mental model is supplied, statistics
are returned for all aggregated groups.

If a mental model object with individual user data is
supplied, statistics are returned per user.
}
\description{
Create overview stats for a mental model
}
