% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{calculate_user_stats}
\alias{calculate_user_stats}
\title{Create overview stats over a single user mental model}
\usage{
calculate_user_stats(edgelist, user)
}
\arguments{
\item{edgelist}{Likely a raw edgelist data frame generated from parse_mtools_csv.}

\item{user}{A user ID as appearing in the User_ID column of the edgelist supplied}
}
\value{
A data frame with the following stats:
- weighted betwenness
- weighted indegree
- weighted outdegree
- weighted total degree
}
\description{
Create overview stats over a single user mental model
}
