% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_models_df}
\alias{example_models_df}
\title{Example data frame from anonymized MTool export}
\format{
An object of class \code{data.frame} with 392 rows and 11 columns.
}
\usage{
example_models_df
}
\description{
An data frame imported from a MTool csv export of mental models of various actors from a real data gathering campaign.
IDs shuffled for anonymization.
}
\keyword{datasets}
