% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_aggregate_el}
\alias{get_aggregate_el}
\title{Get a aggregate edgelist from M-Tool Data}
\usage{
get_aggregate_el(edgelist)
}
\arguments{
\item{edgelist}{Likely a raw edgelist data frame generated from parse_mtools_csv.}
}
\value{
A edgelist (data frame) with three columns: To, From, Weight
}
\description{
This function returns a edgelist of aggregate weights
(the simple sum function) given an M-Tool dataframe supplied
}
