% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{get_model_sims}
\alias{get_model_sims}
\title{Get similiarities between mental models of users}
\usage{
get_model_sims(mentalmodel, method = NULL, group_var = NULL, output = "matrix")
}
\arguments{
\item{mentalmodel}{A mtoolr object with models from multiple users (therefore not an aggregated object)}

\item{method}{The method to calculate similarity between mental models of users.
At the moment, this defaults to "gower". Other options are "simple_overlap" and "jaccard".}

\item{group_var}{A character string giving the name of a (preferably factor) variable in the user data to create
groups of users. For each group of users, a similarity matrix is calculated. All matrices
are returned in a list.}

\item{output}{Either "matrix" or "tibble". If "matrix, a similarity matrix is returned.
If "tibble" a tibble dataframe of pairs of users with similarities is returned. Defaults
to "matrix"}
}
\value{
If no group_var is supplied a similarity matrix or tibble, depending on
output chosen. If a group_var is supplied a list of similarity matrices for each level of the grouping variable.
}
\description{
This function calculates the similarity between the mental models of
of users, either among all users or among groups of users.
}
\examples{
# get gower similarity between mental models of all users
gower_sim_mat <- get_model_sims(example_models)
gower_sim_mat
# simulate user data to add
user_df <- data.frame(id = example_models$user_data$id,var = rnorm(length(example_models$user_data$id)))
user_df$group <- ifelse(user_df$var > 0, "group1","group2")
# add user data
example_models <- example_models |> add_user_data(user_data = user_df,id_key = "id")
# compute similarities by group
gower_sim_mat_list <- get_model_sims(example_models, group_var = "group")
}
