% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_user_el}
\alias{get_user_el}
\title{Get the edgelist of a specific user of M-Tool}
\usage{
get_user_el(edgelist, user)
}
\arguments{
\item{edgelist}{Likely a raw edgelist data frame generated from parse_mtools_csv.}

\item{user}{A user ID as appearing in the User_ID column of the edgelist supplied}
}
\value{
A subset edgelist
}
\description{
Get the edgelist of a specific user of M-Tool
}
