% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{get_user_graph}
\alias{get_user_graph}
\title{Get the igraph graph object for a given user}
\usage{
get_user_graph(user, x)
}
\arguments{
\item{user}{The id of the user as stored in the User_ID column of the M-Tool output.}

\item{x}{A mtoolr object}
}
\value{

}
\description{
Get the igraph graph object for a given user
}
