% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{get_user_model_sim}
\alias{get_user_model_sim}
\title{Get the similarity between two specific users}
\usage{
get_user_model_sim(user1, user2, mentalmodel, method)
}
\arguments{
\item{user1}{A character string for user 1}

\item{user2}{A character string for user 2}

\item{mentalmodel}{A mtoolr object}

\item{method}{The method to use (see `?get_model_sims` for available)}
}
\value{
A numeric value between 0 and 1
}
\description{
Get the similarity between two specific users
}
