% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{get_users_sim_df}
\alias{get_users_sim_df}
\title{Get an tibble of similarities between all of a group of users}
\usage{
get_users_sim_df(users, mentalmodel, method)
}
\arguments{
\item{users}{A character vector of user IDs}

\item{mentalmodel}{A mtoolr object}

\item{method}{The method to use (see `?get_model_sims` for available)}
}
\value{
A tibble of all user pairs with their similarities and the method used
}
\description{
Get an tibble of similarities between all of a group of users
}
