% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{get_users_sim_mat}
\alias{get_users_sim_mat}
\title{Get the similarity matrix between all of a group of users}
\usage{
get_users_sim_mat(users, mentalmodel, method)
}
\arguments{
\item{users}{A character vector of user IDs}

\item{mentalmodel}{A mtoolr object}

\item{method}{The method to use (see `?get_model_sims` for available)}
}
\value{
A similarity matrix
}
\description{
Get the similarity matrix between all of a group of users
}
