% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{igraph_from_mtools_el}
\alias{igraph_from_mtools_el}
\title{Create an igraph graph object from a M-Tool edgelist}
\usage{
igraph_from_mtools_el(edgelist)
}
\arguments{
\item{edgelist}{Likely a raw edgelist data frame generated from parse_mtools_csv.}
}
\value{
A weighted, directed igraph object
}
\description{
Create a weighted, direct igraph graph object from M-Tool data.
This makes it easy to process the data further using various network
statistics
}
