% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{is_mtoolr}
\alias{is_mtoolr}
\title{Is the object of class mtoolr?}
\usage{
is_mtoolr(object)
}
\arguments{
\item{object}{An R object}
}
\value{
TRUE if the object is of class mtoolr
}
\description{
Is the object of class mtoolr?
}
\examples{
is_mtoolr(example_models)
is_mtoolr(mtcars)
}
