% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mental_model_ggraph}
\alias{mental_model_ggraph}
\title{A ggraph plotting setup for mental model data}
\usage{
mental_model_ggraph(graph)
}
\arguments{
\item{graph}{Igraph object based on mental model data}
}
\value{
A ggplot object
}
\description{
A ggraph plotting setup for mental model data
}
