% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{mentalmodel}
\alias{mentalmodel}
\title{Create a mtoolr object}
\usage{
mentalmodel(x, aggregated = FALSE)
}
\arguments{
\item{x}{An object to be converted to a mtoolr object for further analysis. Likely a data frame
with at least User_ID, To, From and Weight columns specifying mental model edges by user.}

\item{aggregated}{Is the object an aggregated model? Defaults to FALSE.}
}
\value{
A object of class mtoolr
}
\description{
This function creates a mtoolr object from a data frame parsed from MTool output.
This function is useful if you want to process exported data from MTool before passing it to MToolR.
The safest and easiest way to import MTool data directly is to use the `mentalmodel_from_csv()`
function however.
}
