% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{mentalmodel_from_csv}
\alias{mentalmodel_from_csv}
\title{Import CSV file exported by M-Tool as mtoolr object}
\usage{
mentalmodel_from_csv(csv_file, exclude_nonresponse = TRUE)
}
\arguments{
\item{csv_file}{Path to a .csv file exported from M-Tool}

\item{exclude_nonresponse}{Should non-responses be excluded? Defaults to TRUE}
}
\value{
An object of class mtoolr containing all information included in the M-Tool export
}
\description{
Import CSV file exported by M-Tool as mtoolr object
}
\examples{
csv_path <- system.file("extdata", "example_export.csv", package = "MToolR", mustWork = TRUE)
mental_model <- mentalmodel_from_csv(csv_path, exclude_nonresponse = TRUE)
mental_model
}
