% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MToolR-package.R
\docType{package}
\name{MToolR-package}
\alias{MToolR}
\alias{MToolR-package}
\title{MToolR: Companion package for the Mental Model Mapping Tool M-Tool. https://www.m-tool.org/}
\description{
This is a companion R package for the Mental Model Mapping Tool M-Tool. https://www.m-tool.org/ The package provides: a) functionality to load and process data generated by M-Tool b) procedures for common analytical tasks c) visualization functions
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marioangst/MToolR}
  \item \url{https://marioangst.github.io/MToolR}
  \item Report bugs at \url{https://github.com/marioangst/MToolR/issues}
}

}
\keyword{internal}
