% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new_mtoolr}
\alias{new_mtoolr}
\title{mtoolr object constructor}
\usage{
new_mtoolr(x = list(), aggregated = logical(), concepts = NULL)
}
\arguments{
\item{x}{A list object passing `is_valid_mtool_edgelist`}

\item{aggregated}{A boolean - is the object aggregated?}

\item{concepts}{An optional list of unique concepts in the mental model. Present to guard against the case of isolates after aggregation.}
}
\value{

}
\description{
This is the internal object constructor for the mtoolr package.
The user should generally not have to engage with it.
}
