% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_mtool_csv}
\alias{parse_mtool_csv}
\title{Parse an M-Tool csv file to a data frame}
\usage{
parse_mtool_csv(csv_file, exclude_nonresponse)
}
\arguments{
\item{csv_file}{Path to a .csv file exported from M-Tool}

\item{exclude_nonresponse}{Should non-responses be excluded? Defaults to TRUE}
}
\value{
A data frame object of your mtool csv
}
\description{
Use this function if you want to edit your M-Tool data before
using it with MToolR. In most cases, you'll want to use
`mentalmodel_from_csv()` directly.
}
\examples{
csv_path <- system.file("extdata", "example_export.csv", package = "MToolR", mustWork = TRUE)
mental_model_df <- parse_mtool_csv(csv_path, exclude_nonresponse = TRUE)
mental_model_df
}
