% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_mtools_csv}
\alias{parse_mtools_csv}
\title{Parse CSV file exported by M-Tool}
\usage{
parse_mtools_csv(csv_file, exclude_nonresponse = TRUE)
}
\arguments{
\item{csv_file}{Path to a .csv file exported from M-Tool}

\item{exclude_nonresponse}{Should non-responses be excluded? Defaults to TRUE}
}
\value{
A data.frame object containing all information included in the M-Tool export
}
\description{
Parse CSV file exported by M-Tool
}
