% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.mtoolr}
\alias{plot.mtoolr}
\title{Default mental model plot function using ggraph}
\usage{
\method{plot}{mtoolr}(mentalmodel, ...)
}
\arguments{
\item{mentalmodel}{An object of type mtoolr}
}
\value{
A plot
}
\description{
Default mental model plot function using ggraph
}
