% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_aggregate_model}
\alias{plot_aggregate_model}
\title{Plot the aggregate model from an edgelist supplied by M-Tool}
\usage{
plot_aggregate_model(edgelist, ...)
}
\arguments{
\item{edgelist}{Likely a raw edgelist data frame generated from parse_mtools_csv().}
}
\value{
A ggplot object
}
\description{
Plot the aggregate model from an edgelist supplied by M-Tool
}
