% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mentalmodel}
\alias{plot_mentalmodel}
\title{Plot mental models using ggraph}
\usage{
plot_mentalmodel(mentalmodel, user = NULL, ...)
}
\arguments{
\item{mentalmodel}{An object of type mtoolr}

\item{user}{A character string giving the name of a specific user for which to plot a mental model}

\item{...}{other parameters to ggraph, eg. layout choice - normally defaults
to "stress"
Check the graphlayouts package for more options, eg. "circle" or "sugiyama"}
}
\value{
A ggplot object
}
\description{
If data on multiple users are supplied, a sample user model is plotted.
If you supply a user argument, you can plot the model of a specific user.
If aggregated data are supplied, the aggregated model is plotted.
}
\examples{
# NOT RUN
# #plot a random user model
# plot_mentalmodel(example_models)
# #plot a specific user
# plot_mentalmodel(example_models, user = "ad84c4ed-b73e-4ba2-8e1f-edbe365bb225")
# #aggregate and then plot
# aggregated_model <- aggregate_mentalmodel(example_models)
# plot_mentalmodel(aggregated_model)
}
