% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mtoolr}
\alias{print.mtoolr}
\title{Default mental model print method}
\usage{
\method{print}{mtoolr}(x, ...)
}
\arguments{
\item{x}{The mental model}

\item{...}{}
}
\value{
Prints a tibble of the raw data and the combined graph object
}
\description{
Default mental model print method
}
