% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{aggregate_mentalmodel}
\alias{aggregate_mentalmodel}
\title{Create aggregate edgelist from M-Tool Data}
\usage{
aggregate_mentalmodel(mentalmodel, aggregate_function = "median")
}
\arguments{
\item{mentalmodel}{A mtoolr object}
}
\value{
A aggregated mtoolr object
}
\description{
This function returns a mental model aggregating all
individual mental models in a mtoolr object.
The current default aggregation function is the median weight.
}
