% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_models}
\alias{example_models}
\title{Example mtoolr class object from anonymized MTool export}
\format{
An object of class \code{mtoolr} of length 5.
}
\usage{
example_models
}
\description{
An object read from a MTool csv export of mental models of various actors from a real data gathering campaign.
IDs shuffled for anonymization.
}
\keyword{datasets}
