% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.mtoolr}
\alias{plot.mtoolr}
\title{Default mental model plot function using ggraph}
\usage{
\method{plot}{mtoolr}(x, ...)
}
\arguments{
\item{x}{An object of type mtoolr}

\item{...}{other parameters to ggraph, eg. layout choice - normally defaults
to "stress"
Check the graphlayouts package for more options, eg. "circle" or "sugiyama"}
}
\value{
A ggplot object
}
\description{
Calls `plot_mentalmodel()`.
If data on multiple users are supplied, a sample user model is plotted.
If you supply a user argument, you can plot the model of a specific user.
If aggregated data are supplied, the aggregated model is plotted.
}
\examples{
# NOT RUN
# #plot a random user model
# plot(example_models)
# #plot a specific user
# plot(example_models, user = "ad84c4ed-b73e-4ba2-8e1f-edbe365bb225")
# #aggregate and then plot
# aggregated_model <- aggregate_mentalmodel(example_models)
# plot(aggregated_model)
}
