% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_user_model}
\alias{plot_user_model}
\title{Plot the mental model of a specific user}
\usage{
plot_user_model(user, mentalmodel, ...)
}
\arguments{
\item{user}{The id of the user}

\item{mentalmodel}{A mtoolr object containing the mental model of the user}
}
\value{
A ggplot object
}
\description{
Plot the mental model of a specific user
}
\examples{
# plot_user_model(example_models, user = "ad84c4ed-b73e-4ba2-8e1f-edbe365bb225")
}
