% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_user_model}
\alias{plot_user_model}
\title{Plot the mental model of a specific user}
\usage{
plot_user_model(user, mentalmodel, ...)
}
\arguments{
\item{user}{A user ID to be displayed}

\item{mentalmodel}{A mtoolr object containing the mental model of the user}
}
\description{
Plot the mental model of a specific user
}
