% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_user_model}
\alias{plot_user_model}
\title{Plot the mental model of a specific user}
\usage{
plot_user_model(edgelist, user)
}
\arguments{
\item{edgelist}{Likely a raw edgelist data frame generated from parse_mtools_csv().}

\item{user}{A user ID as appearing in the User_ID column of the edgelist supplied}
}
\description{
Plot the mental model of a specific user
}
